PaymentId in Order is now long? instead of int?

+++ DTShop.OrderService.Data/Entities/Order.cs
@@ -9,7 +9,7 @@ namespace DTShop.OrderService.Data.Entities
         public int OrderId { get; set; }
         [Required]
         public string Username { get; set; }
-        public int? PaymentId { get; set; }
+        public long? PaymentId { get; set; }
         public OrderStatus Status { get; set; }
         public ICollection<OrderItem> OrderItems { get; set; }
     }
